pyproject.toml
[project]
name = "pydepgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "code2flow>=2.5.1",
    "kuzu>=0.11.0",
    "tach>=0.29.0",
]

[dependency-groups]
dev = [
    "pytest-cov>=6.2.1",
    "pytest-sugar>=1.0.0",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.12.4",
]

[tool.setuptools.packages.find]
where = ["src"]
